change how the card works

+++ modules/message.py
@@ -10,16 +10,19 @@ class MessageModule:
         self.config = get_config()
 
     async def init_gif(self):
+        regex = (
+            fr'(?:view\/)?({re.escape("trollszn123-ronaldo-gif-18268194")}'
+            fr'|{re.escape("boOys" + ".gif")})'
+        )
         for snowflake in self.config["guilds"]:
-            if self.event.guild.id == snowflake:
-                regex = (
-                    fr'(?:view\/)?({re.escape("trollszn123-ronaldo-gif-18268194")}'
-                    fr'|{re.escape("boOys" + ".gif")})'
-                )
-                if bool(re.search(re.escape("https://tenor.com/") + regex, self.event.content)):
-                    await self.event.delete()
+            if self.event.guild.id == snowflake and bool(
+                    re.search(re.escape("https://tenor.com/") + regex, self.event.content)):
+                await self.event.delete()
             elif self.event.guild.id == snowflake and self.config["owner_id"] == self.event.author.id:
-                if bool(re.match(r"^https:\/\/tenor\.com\/(?:view\/)?[a-zA-Z0-9_-]+(?:\.gif)?$", self.event.content)):
+                if bool(re.match(
+                        r"^https:\/\/tenor\.com\/(?:view\/)?[\w\d\u0080-\uFFFF%-]+(?:\.gif)?$",
+                        self.event.content
+                )):
                     if self.event.reference:
                         await self.event.channel.send(content=self.event.content, reference=self.event.reference)
                     else:
